sha adapted for fisoc + regress scipts versioned now

+++ sha/sha.h
@@ -7,7 +7,7 @@
 /* Useful defines & typedefs */
 
 typedef unsigned char BYTE;
-typedef unsigned long LONG;
+typedef unsigned long long LONG;
 
 #define SHA_BLOCKSIZE		64
 #define SHA_DIGESTSIZE		20
